+++ test/test_gridfs_bucket.py
@@ -86,10 +86,8 @@ class TestGridfs(IntegrationTest):
             cls.db, bucket_name="alt")
 
     def setUp(self):
-        self.db.drop_collection("fs.files")
-        self.db.drop_collection("fs.chunks")
-        self.db.drop_collection("alt.files")
-        self.db.drop_collection("alt.chunks")
+        self.cleanup_colls(self.db.fs.files, self.db.fs.chunks,
+                           self.db.alt.files, self.db.alt.chunks)
 
     def test_basic(self):
         oid = self.fs.upload_from_stream("test_filename",
@@ -105,7 +103,6 @@ class TestGridfs(IntegrationTest):
         self.assertEqual(0, self.db.fs.chunks.count_documents({}))
 
     def test_multi_chunk_delete(self):
-        self.db.fs.drop()
         self.assertEqual(0, self.db.fs.files.count_documents({}))
         self.assertEqual(0, self.db.fs.chunks.count_documents({}))
         gfs = gridfs.GridFSBucket(self.db)
@@ -130,7 +127,7 @@ class TestGridfs(IntegrationTest):
         self.assertEqual(oid, raw["_id"])
         self.assertTrue(isinstance(raw["uploadDate"], datetime.datetime))
         self.assertEqual(255 * 1024, raw["chunkSize"])
-        self.assertTrue(isinstance(raw["md5"], str))
+        self.assertNotIn("md5", raw)
 
     def test_corrupt_chunk(self):
         files_id = self.fs.upload_from_stream("test_filename",
@@ -147,12 +144,11 @@ class TestGridfs(IntegrationTest):
             self.fs.delete(files_id)
 
     def test_upload_ensures_index(self):
-        # setUp has dropped collections.
-        names = self.db.list_collection_names()
-        self.assertFalse([name for name in names if name.startswith('fs')])
-
         chunks = self.db.fs.chunks
         files = self.db.fs.files
+        # Ensure the collections are removed.
+        chunks.drop()
+        files.drop()
         self.fs.upload_from_stream("filename", b"junk")
 
         self.assertTrue(any(
@@ -464,41 +460,20 @@ class TestGridfs(IntegrationTest):
         self.assertEqual(file1.read(), file2.read())
 
     def test_md5(self):
-        gin = self.fs.open_upload_stream("has md5")
-        gin.write(b"includes md5 sum")
-        gin.close()
-        self.assertIsNotNone(gin.md5)
-        md5sum = gin.md5
-
-        gout = self.fs.open_download_stream(gin._id)
-        self.assertIsNotNone(gout.md5)
-        self.assertEqual(md5sum, gout.md5)
-
-        gin = self.fs.open_upload_stream_with_id(ObjectId(), "also has md5")
-        gin.write(b"also includes md5 sum")
-        gin.close()
-        self.assertIsNotNone(gin.md5)
-        md5sum = gin.md5
-
-        gout = self.fs.open_download_stream(gin._id)
-        self.assertIsNotNone(gout.md5)
-        self.assertEqual(md5sum, gout.md5)
-
-        fs = gridfs.GridFSBucket(self.db, disable_md5=True)
-        gin = fs.open_upload_stream("no md5")
+        gin = self.fs.open_upload_stream("no md5")
         gin.write(b"no md5 sum")
         gin.close()
         self.assertIsNone(gin.md5)
 
-        gout = fs.open_download_stream(gin._id)
+        gout = self.fs.open_download_stream(gin._id)
         self.assertIsNone(gout.md5)
 
-        gin = fs.open_upload_stream_with_id(ObjectId(), "also no md5")
+        gin = self.fs.open_upload_stream_with_id(ObjectId(), "also no md5")
         gin.write(b"also no md5 sum")
         gin.close()
         self.assertIsNone(gin.md5)
 
-        gout = fs.open_download_stream(gin._id)
+        gout = self.fs.open_download_stream(gin._id)
         self.assertIsNone(gout.md5)